requirements.lock.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
datasets==4.0.0
dill==0.3.8
distro==1.9.0
dnspython==2.8.0
faiss-cpu==1.13.2
fastapi==0.136.0
filelock==3.28.0
frozenlist==1.8.0
fsspec==2025.3.0
ftfy==6.3.1
groq==1.1.2
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
Markdown==3.10.2
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.16
networkx==3.6.1
numpy==2.4.4
open_clip_torch==2.32.0
packaging==26.1
pandas==3.0.2
pillow==11.3.0
pluggy==1.6.0
propcache==0.4.1
protobuf==7.34.1
pyarrow==19.0.1
pydantic==2.13.1
pydantic_core==2.46.1
Pygments==2.20.0
pymongo==4.16.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
safetensors==0.7.0
six==1.17.0
sniffio==1.3.1
starlette==1.0.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
timm==1.0.26
tokenizers==0.22.2
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
wcwidth==0.6.0
Werkzeug==3.1.8
xxhash==3.6.0
yarl==1.23.0
